+++ platform/exacto/commander/exacto_data_storage.h
@@ -123,7 +123,7 @@ extern uint8_t setupReceiveLengthExactoDataStorage( const uint8_t length);
 extern uint8_t clearExactoDataStorage();
 extern thread_control_result_t getStateExactoDataStorage();
 extern uint8_t setDataToExactoDataStorage(uint8_t * data, const uint16_t datacount, thread_control_result_t result);
-extern uint8_t getMailFromExactoDataStorage(uint8_t * receiver, const uint8_t receiver_length);
+extern uint8_t getMailFromExactoDataStorage(uint8_t * receiver, const uint16_t receiver_length);
 
 extern exactolink_package_result_t ex_checkData_ExDtStr();
 extern uint16_t ex_getData_ExDtStr(uint8_t * buffer, uint16_t buffer_length, uint16_t data_type);
